Commit for Vi-Long

+++ httpd/src/http/request.h
@@ -9,7 +9,7 @@
 int extract_method(struct request *req, const char *raw_request, size_t len);
 int extract_target(struct request *req, const char *raw_request, size_t len);
 int extract_version(struct request *req, const char *raw_request, size_t len);
-int extract_headers(struct request *req, const char *raw_request, size_t len);
+int extract_headers(struct request *req, const char *raw_request, int len);
 int check_method(enum method method);
 int check_version(struct string *version);
 int check_missing_header(struct header *headers, struct config *config);
